Revert "Represent the location as an truple of Integers"
This reverts commit 5a156d46d4d327be4ab2a8912318567baaac5233.

+++ src/main/java/com/imagefuzzy/data/Region.java
@@ -7,7 +7,7 @@ import java.awt.image.BufferedImage;
  *
  * @author @author Néstor Rodríguez Vico (nrv23@correo.ugr.es).
  */
-public class Region extends Tuple<BufferedImage, Tuple<Integer, Integer>> {
+public class Region extends Tuple<BufferedImage, Tuple<Double, Double>> {
 
     /**
      * Construct a region.
@@ -15,7 +15,7 @@ public class Region extends Tuple<BufferedImage, Tuple<Integer, Integer>> {
      * @param image    image of the region.
      * @param location location of the region.
      */
-    public Region(BufferedImage image, Tuple<Integer, Integer> location) {
+    public Region(BufferedImage image, Tuple<Double, Double> location) {
         super(image, location);
     }
 
@@ -33,7 +33,7 @@ public class Region extends Tuple<BufferedImage, Tuple<Integer, Integer>> {
      *
      * @return the location of the region.
      */
-    public Tuple<Integer, Integer> getLocation() {
+    public Tuple<Double, Double> getLocation() {
         return this.getSecond();
     }
 }
